# Conflicts:
#	rllib/BUILD
#	rllib/algorithms/algorithm.py
#	rllib/algorithms/algorithm_config.py
#	rllib/algorithms/appo/tf/appo_tf_learner.py
#	rllib/algorithms/appo/torch/appo_torch_learner.py
#	rllib/algorithms/impala/impala.py
#	rllib/algorithms/impala/tf/impala_tf_learner.py
#	rllib/algorithms/impala/torch/impala_torch_learner.py
#	rllib/algorithms/ppo/ppo.py
#	rllib/algorithms/ppo/ppo_learner.py
#	rllib/algorithms/ppo/tf/ppo_tf_rl_module.py
#	rllib/algorithms/ppo/torch/ppo_torch_learner.py
#	rllib/algorithms/ppo/torch/ppo_torch_rl_module.py
#	rllib/connectors/connector_pipeline_v2.py
#	rllib/connectors/learner/default_learner_connector.py
#	rllib/core/learner/learner.py
#	rllib/core/learner/learner_group.py
#	rllib/env/single_agent_env_runner.py
#	rllib/evaluation/postprocessing_v2.py
#	rllib/evaluation/worker_set.py
#	rllib/policy/sample_batch.py
#	rllib/tuned_examples/ppo/pendulum_ppo_envrunner.py
#	rllib/utils/actor_manager.py
#	rllib/utils/minibatch_utils.py
#	rllib/utils/postprocessing/episodes.py
#	rllib/utils/postprocessing/tests/test_value_predictions.py
#	rllib/utils/postprocessing/value_predictions.py
#	rllib/utils/postprocessing/zero_padding.py

+++ rllib/utils/postprocessing/value_predictions.py
@@ -32,7 +32,7 @@ def compute_value_targets(
     # Reverse back to correct (time) direction.
     value_targets = np.stack(list(reversed(Rs)), axis=0)
 
-    return value_targets
+    return value_targets.astype(np.float32)
 
 
 def extract_bootstrapped_values(vf_preds, episode_lengths, T):
@@ -91,7 +91,7 @@ def extract_bootstrapped_values(vf_preds, episode_lengths, T):
         # - Skip the additional timestep at the end and ,ove on with next episode.
         elif T == eps_len:
             bootstrapped_values.append(vf_preds[T])
-            vf_preds = vf_preds[T + 1:]
+            vf_preds = vf_preds[T + 1 :]
         # The episode fits entirely into the T-stride ->
         # - Move on to next episode ("fix" its length by make it seemingly longer).
         else:
